ion-c now builds inside $OUT_DIR
Cargo requires that build scripts only modify the contents of
the $OUT_DIR folder, a path which is passed in as an environment
variable.

Previously, the build script would create a new folder in the
ion-c submodule and store its build artifacts there. This caused
Cargo to object when it tried to package the crate for publishing.