Enable `inspect` to display symbol table encodings
`ion-rs` v0.8.0 introduced a `SystemReader` which allows applications to
inspect local symbol tables in the stream with the same granularity with
which they can inspect user-level Ion values.

This PR upgrades the `beta inspect` command to use the new
`SystemReader` instead of the now-defunct `SystemEventHandler` callback
API. This allows `inspect` to show the complete contents of any symbol
table it encounters.