Makes `to` and `from` into namespaces with a nested `json` command (#54)
Prior to this PR, `to` and `from` were commands that each took a `format`
flag that could be used to specify `json`. This patch promotes `to` and
`from` into namespaces that each have a nested `json` command. This makes
it easier to add support for other subcommands in `to` and `from`, and
allows those subcommands to expose settings that only make sense for the
corresponding format. (For example: configuring a catalog for Ion data,
or a list of supported tags for CBOR.)

Also fixes #36 by manually outputting a newline after writing any file
in an Ion text format.